rename serde with validator
not using serde json or core
